simplify the ex `erigon` example
add helper methods to track active and starting instances
